pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "influx"
version = "0.1.0"
description = "Influx"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12,<3.13"
authors = [{ name = "Dave Snowdon" }]
dependencies = [
    "httpx>=0.27,<1",
    "mcp>=1.8,<2",
    "python-dotenv>=1.0",
    "pydantic>=2.0,<3",
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "apscheduler>=3.10,<4",
    "trafilatura>=2.0",
    "pypdf>=5.0",
    "feedparser>=6.0,<7",
]

[project.optional-dependencies]
otel = [
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp-proto-http>=1.20",
]

[project.scripts]
influx = "influx.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/influx"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pyright>=1.1",
    "respx>=0.22",
    "ruff>=0.4",
]

[tool.ruff]
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pyright]
pythonVersion = "3.12"
venvPath = "."
venv = ".venv"
typeCheckingMode = "standard"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
